- Marten added support for some more ANSI control sequences in csPrintf on windows (where they are emulated). - Marten fixed an ambiguous function call in pagingformer.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@23803 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/terraformer/pagingformer/pagingformer.cpp
@@ -504,8 +504,8 @@ printf("!bigger\n");
           tostep = resolution;
 
           // to this former's relative position to the whole request
-          tox = (uint)ceil( (k/xcount) * partialresolution);  // can't replace by using scale+offset
-          toy = (uint)ceil( (k%xcount) * partialresolution);  // because can't get sampler's former
+          tox = (uint)ceilf( (k/xcount) * partialresolution);  // can't replace by using scale+offset
+          toy = (uint)ceilf( (k%xcount) * partialresolution);  // because can't get sampler's former
         }
 
 /*
